// payroll-web/src/pages/payroll/PayrollDetailPage.types.ts
export interface ProcessingRow {
  nameId: string;
  employeeCode: string;
  firstName: string;
  lastName: string;
  groupId: string;
  positionId: string;
  areaId: string;
  daysWorked: number;
  absences: number;
  lateHours: number;
  overtimeHours: number;
  basicSalary: number;
  ratePerDay: number;
  salaryAmount: number;
}